fixed bugs, added a config class
Config class now saves and loads values based on carId (key), currently it will NOT load your values for cars at the main menu I am attempting to fix this!

TODO: I'd like to find where the dynostand value is at and have that adjusted at the same time the BTM value is adjusted! as well as save that value too.